Fix BasicClusterTest.Observe test case on s390x
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170329
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/ep_engine.cc
@@ -5279,7 +5279,7 @@ cb::engine_errc EventuallyPersistentEngine::observe(
         }
 
         // Get key stats
-        uint16_t keystatus = 0;
+        uint8_t keystatus = 0;
         struct key_stats kstats = {};
         rv = kvBucket->getKeyStats(
                 key, vb_id, cookie, kstats, WantsDeleted::Yes);
